+++ example/widgets/lib/app/modules/chat/views/chatdetail_view.dart
@@ -16,7 +16,7 @@ class ChatdetailView extends StatelessWidget {
     return Scaffold(
       body: AppService.widgets.chat.chatdetailWidget(
         context,
-        chat: controller.chat.value!,
+        cachedChat: controller.chat.value!,
         chatImage:
             "https://images.pexels.com/photos/2486168/pexels-photo-2486168.jpeg?auto=compress&cs=tinysrgb&dpr=1&w=500",
         chatcall: (chat) {